requirements.txt
faiss-cpu
sentence-transformers
llama-cpp-python
numpy
torch
transformers
jupyterlab  # If running in Jupyter Notebook